feat: Separate Asset Container
